move: 메인페이지 경로 이동
